hhalign.C - add -dgrayscale option
This option enables grayscale dotplot when -norealign mode
is used. When it is enabled, -dthr threshold is used as a scaling factor, it
represents the maximum value (i.e. dots which are >= this value are black).

By default it is off and the dotplot is black and white (-dthr is used as a threshold),
as it was in the original 2.0.16 version.